[POC] Unify tracking and testing code UI
